feat: useInputs 훅 구현
