+++ src/components/domain/CardNumbersField.tsx
@@ -8,9 +8,10 @@ import {
   DINERS_CARD_NUMBERS_LENGTH,
   ERROR_MESSAGES,
 } from '../../constants';
-import { useEffect, useEffectEvent, useRef } from 'react';
+import { useEffect, useEffectEvent } from 'react';
 import { sanitizeNumber } from '../../utils';
 import { validates } from '../../validates.ts';
+import { useInputs } from '../../hooks/useInputs.ts';
 
 interface CardNumbersFieldProps {
   value: CardInfo['cardNumbers'];
@@ -29,6 +30,8 @@ export default function CardNumbersField({
   setFieldError,
   onCompleted,
 }: CardNumbersFieldProps) {
+  const { inputRefs, moveToNext, handleKeyDown } = useInputs();
+
   const cardNumbersLength = (() => {
     if (cardBrand === 'diners') return DINERS_CARD_NUMBERS_LENGTH;
     if (cardBrand === 'amex') return AMEX_CARD_NUMBERS_LENGTH;
@@ -63,8 +66,6 @@ export default function CardNumbersField({
     return null;
   };
 
-  const inputRefs = useRef<HTMLInputElement[]>([]);
-
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>, index: number) => {
     const inputValue = e.target.value;
     const error = validate('change', inputValue, index);
@@ -73,22 +74,7 @@ export default function CardNumbersField({
     setFieldValue('cardNumbers', sanitizeNumber(inputValue), index);
 
     if (inputValue.length === cardNumbersLength[index]) {
-      inputRefs.current[index + 1]?.focus();
-    }
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>, index: number) => {
-    if (value[index] === '' && e.key === 'Backspace') {
-      e.preventDefault();
-      inputRefs.current[index - 1]?.focus();
-    }
-
-    if (inputRefs.current[index]?.selectionStart === 0 && e.key === 'ArrowLeft') {
-      inputRefs.current[index - 1]?.focus();
-    }
-
-    if (inputRefs.current[index]?.selectionStart === value[index].length && e.key === 'ArrowRight') {
-      inputRefs.current[index + 1]?.focus();
+      moveToNext(index);
     }
   };
